Change dstrlen to return size_t

+++ ft_dstrlen.c
@@ -6,15 +6,15 @@
 /*   By: abarnett <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/26 12:19:17 by abarnett          #+#    #+#             */
-/*   Updated: 2018/05/31 15:20:57 by abarnett         ###   ########.fr       */
+/*   Updated: 2019/03/02 21:35:32 by alan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_dstrlen(const char *str, char del)
+size_t	ft_dstrlen(const char *str, char del)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] && str[i] != del)
