+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: abarnett <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/17 13:56:25 by abarnett          #+#    #+#             */
-/*   Updated: 2018/06/07 20:40:58 by abarnett         ###   ########.fr       */
+/*   Updated: 2019/03/02 21:33:54 by alan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ char				*ft_strsub(const char *s, unsigned int start, size_t len);
 char				*ft_strjoin(const char *s1, const char *s2);
 char				*ft_strtrim(const char *s);
 int					ft_count_words(const char *str, char del);
-int					ft_dstrlen(const char *str, char del);
+size_t				ft_dstrlen(const char *str, char del);
 char				**ft_strsplit(const char *s, char c);
 int					ft_numlen(long num);
 int					ft_numlen_base(long num, int base);
